Rename pqa button to pqa
