+++ packages/klevu-ui-react/src/components/stencil-generated/index.ts
@@ -36,7 +36,7 @@ export const KlevuPopup = /*@__PURE__*/createReactComponent<JSX.KlevuPopup, HTML
 export const KlevuProduct = /*@__PURE__*/createReactComponent<JSX.KlevuProduct, HTMLKlevuProductElement>('klevu-product');
 export const KlevuProductGrid = /*@__PURE__*/createReactComponent<JSX.KlevuProductGrid, HTMLKlevuProductGridElement>('klevu-product-grid');
 export const KlevuProductQuery = /*@__PURE__*/createReactComponent<JSX.KlevuProductQuery, HTMLKlevuProductQueryElement>('klevu-product-query');
-export const KlevuProductQueryButton = /*@__PURE__*/createReactComponent<JSX.KlevuProductQueryButton, HTMLKlevuProductQueryButtonElement>('klevu-product-query-button');
+export const KlevuProductQueryPopup = /*@__PURE__*/createReactComponent<JSX.KlevuProductQueryPopup, HTMLKlevuProductQueryPopupElement>('klevu-product-query-popup');
 export const KlevuQuery = /*@__PURE__*/createReactComponent<JSX.KlevuQuery, HTMLKlevuQueryElement>('klevu-query');
 export const KlevuQuicksearch = /*@__PURE__*/createReactComponent<JSX.KlevuQuicksearch, HTMLKlevuQuicksearchElement>('klevu-quicksearch');
 export const KlevuRecommendations = /*@__PURE__*/createReactComponent<JSX.KlevuRecommendations, HTMLKlevuRecommendationsElement>('klevu-recommendations');
