+++ packages/klevu-ui/src/components/klevu-chat-messages/readme.md
@@ -11,7 +11,7 @@
 | ----------------------- | ------------------------- | --------------------------------------------------- | ---------------------------------------------------------------------------- | ----------- |
 | `enableMessageFeedback` | `enable-message-feedback` | Should display a feedback button after each message | `boolean \| undefined`                                                       | `undefined` |
 | `feedbacks`             | --                        | Feedbacks given by user                             | `MoiSavedFeedback[] \| undefined`                                            | `undefined` |
-| `messages`              | --                        | Messages received from Moi backend                  | `(MoiResponseText \| MoiResponseFilter \| MoiProducts \| MoiLocalMessage)[]` | `[]`        |
+| `messages`              | --                        | Messages received from Moi backend                  | `(MoiResponseFilter \| MoiResponseText \| MoiProducts \| MoiLocalMessage)[]` | `[]`        |
 | `showFeedbackFor`       | `show-feedback-for`       | What message should we                              | `string \| undefined`                                                        | `undefined` |
 
 
@@ -37,7 +37,7 @@
 ### Used by
 
  - [klevu-moi](../klevu-moi)
- - [klevu-product-query](../klevu-product-query)
+ - [klevu-product-query-popup](../klevu-product-query-popup)
 
 ### Depends on
 
@@ -67,7 +67,7 @@ graph TD;
   klevu-slides --> klevu-util-scrollbars
   klevu-product --> klevu-typography
   klevu-moi --> klevu-chat-messages
-  klevu-product-query --> klevu-chat-messages
+  klevu-product-query-popup --> klevu-chat-messages
   style klevu-chat-messages fill:#f9f,stroke:#333,stroke-width:4px
 ```
