+++ packages/klevu-ui/src/components/klevu-loading-indicator/readme.md
@@ -14,13 +14,13 @@ Simple loading indicator
 ### Used by
 
  - [klevu-moi](../klevu-moi)
- - [klevu-product-query](../klevu-product-query)
+ - [klevu-product-query-popup](../klevu-product-query-popup)
 
 ### Graph
 ```mermaid
 graph TD;
   klevu-moi --> klevu-loading-indicator
-  klevu-product-query --> klevu-loading-indicator
+  klevu-product-query-popup --> klevu-loading-indicator
   style klevu-loading-indicator fill:#f9f,stroke:#333,stroke-width:4px
 ```
