--- packages/klevu-ui/src/components/klevu-product-query-button/klevu-product-query-button.css
@@ -1,3 +1,0 @@
-:host {
-  display: block;
-}
