+++ packages/klevu-ui/src/components/klevu-product-query/klevu-product-query.css
@@ -1,94 +1,3 @@
 :host {
   display: block;
 }
-
-.content {
-  padding: var(--klevu-spacing-04);
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
-
-klevu-chat-layout {
-  height: 500px;
-}
-
-@media (max-width: 520px) {
-  klevu-chat-layout {
-    height: 100%;
-  }
-}
-
-klevu-chat-messages {
-  padding-right: var(--klevu-spacing-04);
-  height: 100%;
-}
-
-.header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.header klevu-icon {
-  cursor: pointer;
-}
-
-div[part="product-query-header"] {
-  width: 100%;
-}
-
-div[part="product-query-footer"] .inputs {
-  display: flex;
-  gap: var(--klevu-spacing-04);
-}
-
-div[part="product-query-footer"] .sendmessage {
-  display: flex;
-  gap: var(--klevu-spacing-04);
-}
-
-div[part="product-query-footer"] .sendmessage klevu-textfield {
-  width: 100%;
-}
-
-.pqa_feedback {
-  text-align: center;
-}
-
-.pqa_feedback > * {
-  padding-bottom: var(--klevu-spacing-04);
-}
-
-.pqa_feedback klevu-icon {
-  font-size: 24px;
-  padding: var(--klevu-spacing-04);
-  cursor: pointer;
-  border-radius: 50%;
-}
-
-.pqa_feedback klevu-icon:hover {
-  background-color: var(--klevu-color-neutral-2);
-}
-
-.pqa_feedback klevu-icon:nth-child(1) {
-  color: var(--klevu-color-accent-1);
-}
-
-.pqa_feedback klevu-icon:nth-child(2) {
-  color: var(--klevu-color-accent-3);
-}
-
-klevu-typography.fineprint {
-  display: block;
-  --klevu-body-xs-size: 10px;
-  padding-bottom: var(--klevu-spacing-04);
-  margin-bottom: var(--klevu-spacing-04);
-  border-bottom: 1px solid var(--klevu-color-neutral-4);
-}
-
-.loading-sorry {
-  display: block;
-  --klevu-body-xs-size: 10px;
-}
