+++ packages/klevu-ui/src/components/klevu-typography/readme.md
@@ -33,7 +33,7 @@ Klevu typography component. This component is used in most places to set correct
  - [klevu-moi](../klevu-moi)
  - [klevu-popular-searches](../klevu-popular-searches)
  - [klevu-product](../klevu-product)
- - [klevu-product-query](../klevu-product-query)
+ - [klevu-product-query-popup](../klevu-product-query-popup)
  - [klevu-quicksearch](../klevu-quicksearch)
  - [klevu-search-landing-page](../klevu-search-landing-page)
  - [klevu-slides](../klevu-slides)
@@ -58,7 +58,7 @@ graph TD;
   klevu-moi --> klevu-typography
   klevu-popular-searches --> klevu-typography
   klevu-product --> klevu-typography
-  klevu-product-query --> klevu-typography
+  klevu-product-query-popup --> klevu-typography
   klevu-quicksearch --> klevu-typography
   klevu-search-landing-page --> klevu-typography
   klevu-slides --> klevu-typography
