+++ packages/klevu-ui/src/global/pqa/pqa_full.css
@@ -107,14 +107,14 @@
   --klevu-elevation-1: rgba(60, 64, 67, 0.3) 0px 1px 2px 0px, rgba(60, 64, 67, 0.15) 0px 2px 6px 2px;
 }
 
-klevu-product-query::part(product-query-footer) {
+klevu-product-query-popup::part(product-query-footer) {
   --klevu-color-primary: #d8edfd;
   --klevu-color-primary-rgb: 216, 237, 253;
   --klevu-color-primary-darker: #b6ccdc;
   --klevu-button-text-color: #000;
 }
 
-klevu-product-query-button {
+klevu-product-query {
   --klevu-color-primary: #428bca;
   --klevu-color-primary-darker: #407db3;
 }
